requirements.txt
opencv-python
matplotlib
numpy
pillow